+++ weldx/core.py
@@ -12,8 +12,8 @@ import xarray as xr
 
 import weldx.util as ut
 from weldx.constants import Q_
-from weldx.constants import WELDX_UNIT_REGISTRY as UREG
-from weldx.time import pandas_time_delta_to_quantity
+from weldx.time import Time, TimeDependent
+from weldx.types import types_time_like
 
 if TYPE_CHECKING:
     import sympy
@@ -254,7 +254,7 @@ class MathematicalExpression:
 # TimeSeries ---------------------------------------------------------------------------
 
 
-class TimeSeries:
+class TimeSeries(TimeDependent):
     """Describes the behaviour of a quantity in time."""
 
     _valid_interpolations = [
@@ -270,8 +270,9 @@ class TimeSeries:
     def __init__(
         self,
         data: Union[pint.Quantity, MathematicalExpression],
-        time: Union[None, pd.TimedeltaIndex, pint.Quantity] = None,
+        time: Union[types_time_like, Time] = None,
         interpolation: str = None,
+        reference_time: pd.Timestamp = None,
     ):
         """Construct a TimSeries.
 
@@ -295,6 +296,7 @@ class TimeSeries:
         self._shape = None
         self._units = None
         self._interp_counter = 0
+        self._reference_time = reference_time
 
         if isinstance(data, (pint.Quantity, xr.DataArray)):
             self._initialize_discrete(data, time, interpolation)
@@ -366,8 +368,8 @@ class TimeSeries:
     def _initialize_discrete(
         self,
         data: Union[pint.Quantity, xr.DataArray],
-        time: Union[None, pd.TimedeltaIndex, pint.Quantity],
-        interpolation: str,
+        time: Union[types_time_like, Time] = None,
+        interpolation: str = None,
     ):
         """Initialize the internal data with discrete values."""
         # set default interpolation
@@ -386,10 +388,16 @@ class TimeSeries:
 
             # constant value case
             if time is None:
-                time = pd.TimedeltaIndex([0])
+                time = pd.Timedelta(0)
 
-            if isinstance(time, pint.Quantity):
-                time = ut.to_pandas_time_index(time)
+            time = Time(time)
+
+            if time.is_absolute:
+                self._reference_time = time.reference_time
+                time = time - time.reference_time
+
+            time = time.as_pandas_index()
+
             if not isinstance(time, pd.TimedeltaIndex):
                 raise ValueError(
                     '"time" must be a time quantity or a "pandas.TimedeltaIndex".'
@@ -440,48 +448,19 @@ class TimeSeries:
                 f' "{str(e)}"'
             )
 
-    def _interp_time_discrete(
-        self, time: Union[pd.TimedeltaIndex, pint.Quantity]
-    ) -> xr.DataArray:
-        """Interpolate the time series if its data is composed of discrete values.
-
-        See `interp_time` for interface description.
-
-        """
-        if isinstance(time, pint.Quantity):
-            time = ut.to_pandas_time_index(time)
-        if not isinstance(time, pd.TimedeltaIndex):
-            raise ValueError(
-                '"time" must be a time quantity or a "pandas.TimedeltaIndex".'
-            )
-
+    def _interp_time_discrete(self, time: Time) -> xr.DataArray:
+        """Interpolate the time series if its data is composed of discrete values."""
         return ut.xr_interp_like(
             self._data,
-            {"time": time},
+            {"time": time.as_timedelta()},
             method=self.interpolation,
             assume_sorted=False,
             broadcast_missing=False,
         )
 
-    def _interp_time_expression(
-        self, time: Union[pd.TimedeltaIndex, pint.Quantity], time_unit: str
-    ) -> xr.DataArray:
-        """Interpolate the time series if its data is a mathematical expression.
-
-        See `interp_time` for interface description.
-
-        """
-        # Transform time to both formats
-        if isinstance(time, pint.Quantity) and time.check(UREG.get_dimensionality("s")):
-            time_q = time
-            time_pd = ut.to_pandas_time_index(time)
-        elif isinstance(time, pd.TimedeltaIndex):
-            time_q = pandas_time_delta_to_quantity(time, time_unit)
-            time_pd = time
-        else:
-            raise ValueError(
-                '"time" must be a time quantity or a "pandas.TimedeltaIndex".'
-            )
+    def _interp_time_expression(self, time: Time, time_unit: str) -> xr.DataArray:
+        """Interpolate the time series if its data is a mathematical expression."""
+        time_q = time.as_quantity(unit=time_unit)
 
         if len(self.shape) > 1 and np.iterable(time_q):
             while len(time_q.shape) < len(self.shape):
@@ -496,7 +475,9 @@ class TimeSeries:
             data = np.expand_dims(data, 0)
 
         dax = xr.DataArray(data=data)  # don't know exact dimensions so far
-        return dax.rename({"dim_0": "time"}).assign_coords({"time": time_pd})
+        return dax.rename({"dim_0": "time"}).assign_coords(
+            {"time": time.as_pandas_index()}
+        )
 
     @property
     def data(self) -> Union[pint.Quantity, MathematicalExpression]:
@@ -569,7 +550,7 @@ class TimeSeries:
         return isinstance(self.data, MathematicalExpression)
 
     @property
-    def time(self) -> Union[None, pd.TimedeltaIndex]:
+    def time(self) -> Union[None, Time]:
         """Return the data's timestamps.
 
         Returns
@@ -579,11 +560,16 @@ class TimeSeries:
 
         """
         if isinstance(self._data, xr.DataArray) and len(self._data.time) > 1:
-            return ut.to_pandas_time_index(self._data.time.data)
+            return Time(self._data.time.data, self.reference_time)
         return None
 
+    @property
+    def reference_time(self) -> Union[pd.Timestamp, None]:
+        """Get the reference time."""
+        return self._reference_time
+
     def interp_time(
-        self, time: Union[pd.TimedeltaIndex, pint.Quantity], time_unit: str = "s"
+        self, time: Union[pd.TimedeltaIndex, pint.Quantity, Time], time_unit: str = "s"
     ) -> "TimeSeries":
         """Interpolate the TimeSeries in time.
 
@@ -594,7 +580,7 @@ class TimeSeries:
         Parameters
         ----------
         time:
-            A set of timestamps.
+            The time values to be used for interpolation.
         time_unit:
             Only important if the time series is described by an expression and a
             'pandas.TimedeltaIndex' is passed to this function. In this case, time is
@@ -615,10 +601,14 @@ class TimeSeries:
                 f"{self._interp_counter} time(s)."
             )
 
+        # prepare timedelta values for internal interpolation
+        time = Time(time)
+        time_interp = Time(time, self.reference_time)
+
         if isinstance(self._data, xr.DataArray):
-            dax = self._interp_time_discrete(time)
+            dax = self._interp_time_discrete(time_interp)
         else:
-            dax = self._interp_time_expression(time, time_unit)
+            dax = self._interp_time_expression(time_interp, time_unit)
 
         ts = TimeSeries(data=dax.data, time=time, interpolation=self.interpolation)
         ts._interp_counter = self._interp_counter + 1
@@ -662,7 +652,7 @@ class TimeSeries:
                 axes=axes, data_name=data_name, time_unit=time_unit, **mpl_kwargs
             )
 
-        time = pandas_time_delta_to_quantity(self.time)
+        time = Time(self.time, self.reference_time).as_quantity()
         if time_unit is not None:
             time = time.to(time_unit)