+++ weldx/tests/test_utility.py
@@ -8,9 +8,6 @@ import pandas as pd
 import pytest
 import xarray as xr
 from numpy import NaN
-from pandas import DatetimeIndex as DTI
-from pandas import TimedeltaIndex as TDI
-from pandas import date_range
 from pint.errors import DimensionalityError
 from xarray import DataArray
 
@@ -117,43 +114,6 @@ def test_vector_is_close():
 
     # vectors have different size
     assert not ut.vector_is_close(vec_a, vec_a[0:2])
-
-
-@pytest.mark.parametrize(
-    "arg, expected",
-    [
-        # timedeltas
-        (TDI([42], unit="ns"), TDI([42], unit="ns")),
-        (pd.timedelta_range("0s", "20s", 10), pd.timedelta_range("0s", "20s", 10)),
-        (np.timedelta64(42), TDI([42], unit="ns")),
-        (np.array([-10, 0, 20]).astype("timedelta64[ns]"), TDI([-10, 0, 20], "ns")),
-        (Q_(42, "ns"), TDI([42], unit="ns")),
-        ("10s", TDI(["10s"])),
-        (["5ms", "10s", "2D"], TDI(["5 ms", "10s", "2D"])),
-        # datetimes
-        (np.datetime64(50, "Y"), DTI(["2020-01-01"])),
-        ("2020-01-01", DTI(["2020-01-01"])),
-        (
-            np.array(
-                ["2012-10-02", "2012-10-05", "2012-10-11"], dtype="datetime64[ns]"
-            ),
-            DTI(["2012-10-02", "2012-10-05", "2012-10-11"]),
-        ),
-    ],
-)
-def test_to_pandas_time_index(arg, expected):
-    """Test conversion to appropriate pd.TimedeltaIndex or pd.DatetimeIndex."""
-    assert np.all(ut.to_pandas_time_index(arg) == expected)
-
-
-@pytest.mark.parametrize(
-    "arg, exception",
-    [(5, TypeError), ("string", TypeError), (Q_(10, "m"), DimensionalityError)],
-)
-def test_to_pandas_time_index_exceptions(arg, exception):
-    """Test correct exceptions on invalid inputs."""
-    with pytest.raises(exception):
-        ut.to_pandas_time_index(arg)
 
 
 class TestXarrayInterpolation:
@@ -354,34 +314,6 @@ def test_xr_interp_like():
         {"t": pd.timedelta_range(start="3s", end="7s", freq="125ms", closed="left")},
     )
     assert np.all(test == np.arange(3, 7, 0.125))
-
-
-@pytest.mark.parametrize(
-    "list_of_objects, time_exp",
-    [
-        (
-            [
-                date_range("2020-02-02", periods=4, freq="2D"),
-                date_range("2020-02-01", periods=4, freq="2D"),
-                date_range("2020-02-03", periods=2, freq="3D"),
-            ],
-            date_range("2020-02-01", periods=8, freq="1D"),
-        ),
-        ([TDI([1, 5]), TDI([2, 6, 7]), TDI([1, 3, 7])], TDI([1, 2, 3, 5, 6, 7])),
-    ],
-)
-def test_get_time_union(list_of_objects, time_exp):
-    """Test input types for get_time_union function.
-
-    Parameters
-    ----------
-    list_of_objects:
-        List with input objects
-    time_exp:
-        Expected result time
-
-    """
-    assert np.all(ut.get_time_union(list_of_objects) == time_exp)
 
 
 def test_xr_fill_all():